server/requirements.txt
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
Flask==3.1.1
gcloud==0.18.3
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
MarkupSafe==3.0.2
numpy==2.2.6
oauth2client==4.1.3
pandas==2.2.3
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pyparsing==3.2.3
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jwt==4.1.0
pytz==2025.2
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9.1
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
Werkzeug==3.1.3
google-auth==2.40.2
google-api-python-client==2.170.0
Flask-Cors==6.0.1
firebase-admin==6.9.0
google-cloud-firestore==2.21.0
google-cloud-functions==1.20.4
